installed threejs, finished solar prototype.
I installed threejs, I believe the packages are in git ignore which is fine, and they got added to package.json.

I also prototyped the solar page. the squares will eventually become solar panels. currently, the sun rotates on page scroll, and there is directional light that changes inversely to the sun.